Delegate repository access to refs, objects
Instead of using the internal field directly to access references
or objects, use the getter method to obtain the proper type of
database, and follow down from there.  This permits us to later
do a refactoring that makes those methods abstract and strips the
field out of the Repository class, moving it into a concrete base
class that is more storage implementation specific.

Change-Id: Ic21dd48800e68a04ce372965ad233485b2a84bef
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>